Add shared abstract class with snap v2 helpers
